pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-wiki"
version = "2.0.0"
description = "LLM-powered Personal Knowledge Base"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "My Wiki Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "loguru>=0.6.0",
    "openai>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
wiki = "wiki.cli:cli"

[project.urls]
Homepage = "https://github.com/Masongmx/my-wiki"
Documentation = "https://github.com/Masongmx/my-wiki#readme"
Repository = "https://github.com/Masongmx/my-wiki"
Issues = "https://github.com/Masongmx/my-wiki/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["wiki*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=wiki --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["wiki"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]